requirements.txt
crewai==0.80.0
crewai-tools==0.14.0
langchain-google-genai==2.0.7
python-dotenv==1.0.1
requests==2.31.0
amadeus==9.0.0
groq==0.13.0
 